requirements.txt
alembic==1.16.5
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.2
colorama==0.4.6
Flask==3.1.1
Flask-Cors==4.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.2.3
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.42
typing_extensions==4.14.1
urllib3==2.5.0
Werkzeug==3.1.3
Stripe==12.5.0
flask-admin==1.6.1
wtforms==3.2.1
gpxpy
polyline